requirements.txt
altgraph==0.15
dis3==0.1.2
future==0.16.0
keyboard==0.13.1
macholib==1.9
mss==3.2.0
pefile==2017.11.5
Pillow==5.1.0
pyaml==17.12.1
PyAutoGUI==0.9.36
PyInstaller==3.3.1
PyMsgBox==1.0.6
PyScreeze==0.1.14
python-xlib==0.23
PyTweening==1.0.3
PyYAML==3.12
six==1.11.0
yapf==0.21.0
